Move some codes from api servert to llumnix_utils
